fix pma

+++ wo/cli/plugins/stack.py
@@ -1005,32 +1005,35 @@ class WOStackController(CementBaseController):
                               .format(WOVariables.wo_webroot))
                     os.makedirs('{0}22222/htdocs/db'
                                 .format(WOVariables.wo_webroot))
-                shutil.move('/tmp/phpmyadmin-STABLE/',
-                            '{0}22222/htdocs/db/pma/'
-                            .format(WOVariables.wo_webroot))
-                shutil.copyfile('{0}22222/htdocs/db/pma/config.sample.inc.php'
-                                .format(WOVariables.wo_webroot),
-                                '{0}22222/htdocs/db/pma/config.inc.php'
+                if not os.path.exists('{0}22222/htdocs/db/'
+                                      'pma/phpmyadmin-STABLE'
+                                      .format(WOVariables.wo_webroot)):
+                    shutil.move('/tmp/phpmyadmin-STABLE/',
+                                '{0}22222/htdocs/db/pma/'
                                 .format(WOVariables.wo_webroot))
-                Log.debug(self, 'Setting Blowfish Secret Key FOR COOKIE AUTH to  '
-                          '{0}22222/htdocs/db/pma/config.inc.php file '
-                          .format(WOVariables.wo_webroot))
-                blowfish_key = ''.join([random.choice
-                                        (string.ascii_letters + string.digits)
-                                        for n in range(25)])
-                WOFileUtils.searchreplace(self,
-                                          '{0}22222/htdocs/db/pma/config.inc.php'
-                                          .format(WOVariables.wo_webroot),
-                                          "$cfg[\'blowfish_secret\'] = \'\';", "$cfg[\'blowfish_secret\'] = \'{0}\';"
-                                          .format(blowfish_key))
-                Log.debug(self, 'Setting HOST Server For Mysql to  '
-                          '{0}22222/htdocs/db/pma/config.inc.php file '
-                          .format(WOVariables.wo_webroot))
-                WOFileUtils.searchreplace(self,
-                                          '{0}22222/htdocs/db/pma/config.inc.php'
-                                          .format(WOVariables.wo_webroot),
-                                          "$cfg[\'Servers\'][$i][\'host\'] = \'localhost\';", "$cfg[\'Servers\'][$i][\'host\'] = \'{0}\';"
-                                          .format(WOVariables.wo_mysql_host))
+                    shutil.copyfile('{0}22222/htdocs/db/pma/config.sample.inc.php'
+                                    .format(WOVariables.wo_webroot),
+                                    '{0}22222/htdocs/db/pma/config.inc.php'
+                                    .format(WOVariables.wo_webroot))
+                    Log.debug(self, 'Setting Blowfish Secret Key FOR COOKIE AUTH to  '
+                              '{0}22222/htdocs/db/pma/config.inc.php file '
+                              .format(WOVariables.wo_webroot))
+                    blowfish_key = ''.join([random.choice
+                                            (string.ascii_letters + string.digits)
+                                            for n in range(25)])
+                    WOFileUtils.searchreplace(self,
+                                              '{0}22222/htdocs/db/pma/config.inc.php'
+                                              .format(WOVariables.wo_webroot),
+                                              "$cfg[\'blowfish_secret\'] = \'\';", "$cfg[\'blowfish_secret\'] = \'{0}\';"
+                                              .format(blowfish_key))
+                    Log.debug(self, 'Setting HOST Server For Mysql to  '
+                              '{0}22222/htdocs/db/pma/config.inc.php file '
+                              .format(WOVariables.wo_webroot))
+                    WOFileUtils.searchreplace(self,
+                                              '{0}22222/htdocs/db/pma/config.inc.php'
+                                              .format(WOVariables.wo_webroot),
+                                              "$cfg[\'Servers\'][$i][\'host\'] = \'localhost\';", "$cfg[\'Servers\'][$i][\'host\'] = \'{0}\';"
+                                              .format(WOVariables.wo_mysql_host))
                 Log.debug(self, 'Setting Privileges of webroot permission to  '
                           '{0}22222/htdocs/db/pma file '
                           .format(WOVariables.wo_webroot))
